Create a separate API class for photos
